# openapi.yaml
openapi: 3.1.0
info:
  title: Verifier API
  description: Generic Verifier service
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /oid4vp/api/request-object/{request_id}/response-data:
    post:
      tags:
      - Verfifier OID4VP API
      summary: Receive Verification Presentation (from e.g. Wallet)
      description: "Handles various types of verification presentations including\
        \ standard presentations, rejections, DCQL presentations, and encrypted DCQL\
        \ presentations. The method automatically determines the request type based\
        \ on the provided parameters."
      externalDocs:
        description: OpenId4VP response parameters
        url: https://openid.net/specs/openid-4-verifiable-presentations-1_0-ID2.html#section-6.1
      operationId: receiveVerificationPresentation
      parameters:
      - name: request_id
        in: path
        description: The unique identifier of the verification request
        required: true
        schema:
          type: string
          format: uuid
      - name: SWIYU-API-Version
        in: header
        description: "Optional API version. Supported values: 2 (DEFAULT) - This supports\
          \ OID4VP 1.0"
        required: false
        schema:
          type: string
          enum:
          - "2"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/VerificationPresentationUnion"
              description: The verification presentation request. Type is determined
                automatically based on the provided parameters.
        required: true
      responses:
        "400":
          description: Bad Request. The request body is not valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "200":
          description: Verification Presentation received and processed successfully
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/VerificationPresentationResponseDto"
        "408":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "410":
          description: Gone. The verification process for this Request Object is already
            closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "415":
          description: Unsupported Media Type. The request body must be application/x-www-form-urlencoded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
...
  /oid4vp/api/request-object/{request_id}:
    get:
      tags:
      - Verfifier OID4VP API
      summary: Get Request Object
      description: Can return a RequestObjectDto as JSON Object or a SignedJWT String
        depending on JAR (JWT secured authorization request) flag in verifier management
      operationId: getRequestObject
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: |-
            Request object either as plaintext or signed JWT.

            The 'application/oauth-authz-req+jwt' representation is a compact serialized JWS (optionally nested JWE) representing the Request Object claims. As this is a JWT and not a JSON object, its structural requirements cannot be expressed as a JSON Schema and are documented here instead:
            - The JOSE header MUST require the 'profile_version' parameter to indicate the Swiss Profile version.
            - The JWT Claims Set corresponds to the [RequestObject](#/components/schemas/RequestObject) schema documented for the 'application/json' representation below, with 'request' and 'request_uri' claims strictly prohibited.

            The 'application/json' representation is kept for documentation purposes only, mirroring the JWT Claims Set of the 'application/oauth-authz-req+jwt' representation; it is not actually returned when JAR (JWT-secured Authorization Request) is enabled.
          content:
            application/oauth-authz-req+jwt:
              schema:
                type: string
                format: jwt
            application/json:
              schema:
                $ref: "#/components/schemas/RequestObject"
        "403":
          description: Forbidden
          content:
            application/oauth-authz-req+jwt:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "404":
          description: Request Object not found
          content:
            application/oauth-authz-req+jwt:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "405":
          description: Method Not Allowed
          content:
            application/oauth-authz-req+jwt:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "406":
          description: Not Acceptable. The requested Accept header does not match
            any of the produced media types
          content:
            application/oauth-authz-req+jwt:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "410":
          description: Gone. The verification process for this Request Object is already
            closed
          content:
            application/oauth-authz-req+jwt:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "500":
          description: Internal Server Error
          content:
            application/oauth-authz-req+jwt:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "503":
          description: Service Unavailable
          content:
            application/oauth-authz-req+jwt:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
  /oid4vp/api/openid-client-metadata.json:
    get:
      tags:
      - Verfifier OID4VP API
      summary: Get client metadata
      description: "Metadata providing further information about the verifier, such\
        \ as name and logo."
      operationId: getOpenIdClientMetadata
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Request object either as plaintext or signed JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenidClientMetadataDto"
              examples:
                Sample Client Metadata:
                  description: Sample Client Metadata
                  value:
                    client_name#en: English name (all regions)
                    client_name#fr: French name (all regions)
                    client_name#de-DE: German name (region Germany)
                    client_name#de-CH: German name (region Switzerland)
                    client_name#de: German name (fallback)
                    client_name: Fallback name
                    client_logo: www.example.com/logo.png
                    client_logo#fr: www.example.com/logo_fr.png
...
components:
  schemas:
    VerificationPresentationUnion:
      type: object
      description: Union DTO that contains all possible parameters from all verification
        presentation types. Only the relevant fields should be populated based on
        the request type.
      properties:
        vp_token:
          description: "VP token that can be either a string for standard presentations\
            \ or a JSON object for DCQL presentations. For standard/PE presentations:\
            \ JWT token string. For DCQL presentations: Object containing credential\
            \ query results where keys are query IDs and values are arrays of presentations."
          example: |-
            Standard/PE: "eyJhbGci...QMA"
            DCQL: {"my_credential": ["eyJhbGci...QMA", "eyJhbGci...QMA"]}
          oneOf:
          - type: string
          - {}
        error:
          type: string
          description: Error code for rejection (used for REJECTION)
          enum:
          - invalid_scope
          - invalid_request
          - invalid_client
          - vp_formats_not_supported
          - invalid_presentation_definition_uri
          - invalid_presentation_definition_reference
          - client_rejected
          - access_denied
          example: client_rejected
        error_description:
          type: string
          description: Error description for rejection (used for REJECTION)
          example: The owner has declined the verification request.
        response:
          type: string
          description: Encrypted response JWE string (used for DCQLE)
          example: eyJhbGciOiJSU0ExXzUiLCJlbmMiOiJBMjU2R0NNIiwidHlwIjoiSldFIn0...
        state:
          type: string
          description: |
            The OAuth State is an opaque value used by the client to maintain state between the request and callback.<br>
            If provided in the request object the state string <em>MUST</em> be returned in the response.
    VerificationPresentationResponseDto:
      type: object
      properties:
        redirect_uri:
          type: string
          format: uri
          description: "[EXPERIMENTAL] Optional: Only provided if set in the initial\
            \ request. It is used to route the response back to the business verifier's\
            \ endpoint."
          example: https://shop.ch/callback?session_nonce=123&response_code=xyz
    ApiErrorDto:
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
        detail:
          type: string
...
    DcqlQueryDto:
      type: object
      description: Represents the Digital Credentials Query Language (DCQL) query
        according to https://openid.net/specs/openid-4-verifiable-presentations-1_0.html#section-6
        **⚠️ Not yet implemented**
      example:
        credentials:
        - id: identity_credential_dcql
          format: dc+sd-jwt
          meta:
            vct_values:
            - https://credentials.example.com/identity_credential
          claims:
          - path:
            - given_name
          - path:
            - family_name
          - path:
            - address
            - street_address
          require_cryptographic_holder_binding: true
        - id: university_degree_dcql
          format: dc+sd-jwt
          meta:
            vct_values:
            - https://credentials.example.com/university_degree
          claims:
          - path:
            - degree_name
          - path:
            - graduation_year
      properties:
        credentials:
          type: array
          description: "A non-empty array of Credential Queries that specify the requested\
            \ Credentials. According to OpenID for Verifiable Presentations 1.0, Section\
            \ 6, property 'credentials'."
          items:
            $ref: "#/components/schemas/DcqlCredentialDto"
          minItems: 1
        credential_sets:
          type: array
          description: "An optional non-empty array of Credential Set Queries that\
            \ specifies additional constraints on which of the requested Credentials\
            \ to return. According to OpenID for Verifiable Presentations 1.0, Section\
            \ 6, property 'credential_sets'."
          items:
            $ref: "#/components/schemas/DcqlCredentialSetDto"
      required:
      - credentials
      x-implementation-status:
        status: not-implemented
...
    ResponseModeType:
      type: string
      description: Supported response_mode as defined in OID4VP
      enum:
      - direct_post
      - direct_post.jwt
...
    JWKSet:
      type: object
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/JsonWebKey"
...
    OpenIdClientMetadataVpFormatsSupported:
      type: object
      properties:
        dc+sd-jwt:
          $ref: "#/components/schemas/OpenIdClientMetadataVpFormatSdJwt"
    OpenidClientMetadataDto:
      type: object
      description: |-
        Verifier metadata values.
        Additional Verifier metadata parameters MAY be defined and used, as described in [RFC7591].
        The Wallet MUST ignore any unrecognized parameters.
      properties:
        client_id:
          type: string
        jwks:
          $ref: "#/components/schemas/JWKSet"
          description: |
            One or more public keys, such as those used by the Wallet as an input to a key agreement
            that may be used for encryption of the Authorization Response, or where the Wallet will
            require the public key of the Verifier to generate a Verifiable Presentation.

            Public keys included in this parameter MUST NOT be used to verify the signature of signed Authorization Requests
        encrypted_response_enc_values_supported:
          type: array
          description: |-
            If present, should be non-empty array of JWE algorithms as in RFC7516.
            When a response_mode requiring encryption of the Response (such as direct_post.jwt) is specified,
            one of the specified algorithms is to be used.
          items:
            type: string
        vp_formats_supported:
          $ref: "#/components/schemas/OpenIdClientMetadataVpFormatsSupported"
          description: |
            An object containing a list of name/value pairs, where the name is a Credential Format Identifier and the
            value defines format-specific parameters that a Verifier supports.
        additionalProperties:
          type: object
          additionalProperties: {}
      required:
      - encrypted_response_enc_values_supported
      - jwks
      - vp_formats_supported
    RequestObject:
      type: object
      description: |
        OID4VP Request Object sent to the Wallet as response after receiving an Authorization Request.
        Contains either a Presentation Exchange (PE) presentation_definition or a DCQL query (dcql_query), depending on the verification request format.
        - If the verification was initiated using the older PE format, the presentation_definition field is used.
        - If the verification was initiated using the new DCQL format, the dcql_query field is used.
      properties:
        client_id:
          type: string
        response_type:
          type: string
        response_mode:
          $ref: "#/components/schemas/ResponseModeType"
          description: |
            If "direct_post", expect the response to be sent via an HTTPS connection to response_uri.
            If "direct_post.jwt" expects the response to also be encrypted.
        response_uri:
          type: string
        nonce:
          type: string
        version:
          type: string
        dcql_query:
          $ref: "#/components/schemas/DcqlQueryDto"
          description: |
            DCQL query object as an Authorization Request parameter.
            This field is used for requests initiated with the DCQL format and contains the Digital Credentials Query Language (DCQL) query.
            Mutually exclusive with the {@code scope} parameter: only one of the two may be present.
        scope:
          type: string
          description: |
            OAuth2 scope value identifying the DCQL query registered in the vqPS.
            MUST be present when a vqPS is injected into verifier_info.
            Mutually exclusive with dcql_query.
        client_metadata:
          $ref: "#/components/schemas/OpenidClientMetadataDto"
          description: "A JSON object containing the Verifier metadata values providing\
            \ further information about the verifier, such as name and logo. It is\
            \ UTF-8 encoded. It MUST NOT be present if client_metadata_uri parameter\
            \ is present."
          example:
            client_name#en: English name (all regions)
            client_name#fr: French name (all regions)
            client_name#de-DE: German name (region Germany)
            client_name#de-CH: German name (region Switzerland)
            client_name#de: German name (fallback)
            client_name: Fallback name
            client_logo: www.example.com/logo.png
            client_logo#fr: www.example.com/logo_fr.png
            vp_formats_supported:
              dc+sd-jwt:
                sd-jwt_alg_values:
                - ES256
                - Ed25519
                kb-jwt_alg_values:
                - ES256
                - Ed25519
        state:
          type: string
          description: |
            An opaque value used by the client to maintain
            state between the request and callback.  The authorization
            server includes this value when redirecting the user-agent back
            to the client.
        aud:
          type: string
          description: |
            Audience ("aud") JWT claim identifying the intended recipient of this Request Object.
            As the verifier cannot identify the wallet the Static Discovery metadata is used.
            MUST be "https://self-issued.me/v2"
        profile_version:
          type: string
          description: JWT header parameter indicating the Swiss Profile version used
            to produce this Request Object.
        encrypted_response_enc_values_supported:
          type: array
          description: JWE 'enc' algorithms accepted by the Verifier for an encrypted
            Authorization Response.
          items:
            type: string
        verifier_info:
          type: array
          description: Trust Protocol 2.0 trust statements embedded as JWT objects.
          items:
            $ref: "#/components/schemas/VerifierInfoEntryDto"
      required:
      - encrypted_response_enc_values_supported
      - profile_version
    VerifierInfoEntryDto:
      type: object
      properties:
        format:
          type: string
        data:
          type: string
...